refactor: reissue 로직 수정
